% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDunedinPACNI.R
\name{ExportDunedinPACNI}
\alias{ExportDunedinPACNI}
\title{ExportDunedinPACNI}
\usage{
ExportDunedinPACNI(data, outdir, missing_ROIs, missing_gwr)
}
\arguments{
\item{data}{Data.frame with the output of LoadFreeSurferStats().}

\item{outdir}{Character string of the path to the desired output directory.}

\item{missing_ROIs}{(optional) Character vector of ROIs that you do not have in your data. ROI names should be formatted according to conventions
    output from LoadFreeSurferStats(). Default is NULL.
    
    IF YOU HAVE MISSING ROIS: this may reduce the accuracy of the output DunedinPACNI scores. Not all ROIs are included in the final model,
    this function will automatically list the ROIs you are missing that will directly impact scores.
    If you are missing ROIs that are included in the DunedinPACNI algorithm, we will automatically impute the mean value from the Dunedin Study
    for that ROI. Each imputed ROI will slightly decrease the accuracy of the overall estimates, so having a lot of missing ROIs reduces 
    confidence in the scores. Also bear in mind, the Dunedin Study is a cohort of 45-year-olds, so missing ROIs (i.e. imputed ROIs) may have
    more dramatic effects in datasets that differ substantially in age from the Dunedin Study members.

    This command will throw an error if you are missing more than >20 percent of the ROIs needed for the DunedinPACNI algorithm}
}
\value{
The output of this function will be a data.frame with DunedinPACNI scores for your dataset. This function will also automatically save this object to an .Rdata file labeled:

      <outdir>/<date and time>_df_pacni.Rdata
}
\description{
This function is used to estimate DunedinPACNI scores from FreeSurfer outputs.
}
\details{
Written by Ethan Whitman (ethan.whitman@duke.edu)
}
\examples{
ExportDunedinPACNI(data = df,
                   outdir = '/Users/ew198/Documents/brainpace/scripts/pacni_package/')

ExportDunedinPACNI(data = df,
                   outdir = '/Users/ew198/Documents/brainpace/scripts/pacni_package/',
                   missing_ROIs = c('GWR_parsopercularis_right', 'GWR_parsorbitalis_right'))
}
