% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFreeSurferStats.R
\name{LoadFreeSurferStats}
\alias{LoadFreeSurferStats}
\title{LoadFreeSurferStats}
\usage{
LoadFreeSurferStats(fsdir, sublistdir, missing_gwr)
}
\arguments{
\item{fsdir}{character string of the directory path where FreeSurfer output files are stored. This directory should contain
the full sample of sub-* directories that contain the stats/ directories. This is likely your FreeSurfer "subjects dir".}

\item{sublistdir}{character string of the directory path where a .csv file of the participant ID list can be found. Participant ID list should
be saved as 'sublist.csv'. This file should have one column with the header/first entry as "ID".}

\item{missing_gwr}{(optional) TRUE/FALSE as to whether gray-white signal intensity ratio phenotypes are unavailable. Default is FALSE.}
}
\value{
The output of this function will be a data.frame with the output of the entire sample's .stats files. Each row
    will have a unique participant and each column will have a different morphometric estimate.
}
\description{
This function is used to quickly read and format FreeSurfer .stats files to R.
}
\details{
Written by Ethan Whitman (ethan.whitman@duke.edu)
}
\examples{
LoadFreeSurferStats(fsdir = '/Users/ew198/Documents/data/freesurfer_stats/',
                    sublistdir = '/Users/ew198/Documents/brainpace/data/')
}
